requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate
opencv-python
pillow
numpy
albumentations
tqdm
matplotlib
scikit-learn
roboflow
python-dotenv
git+https://github.com/UX-Decoder/Segment-Everything-Everywhere-All-At-Once.git@main
